+++ src/input/mms.h
@@ -17,7 +17,7 @@
  * along with this program; if not, write to the Free Software
  * Foundation, Inc., 59 Temple Place - Suite 330, Boston, MA  02111-1307, USA
  *
- * $Id: mms.h,v 1.11 2004/04/06 00:25:29 tmattern Exp $
+ * $Id: mms.h,v 1.12 2006/06/20 01:46:41 dgp85 Exp $
  *
  * libmms public header
  */
@@ -37,7 +37,7 @@ int      mms_read (mms_t *this, char *data, int len);
 uint32_t mms_get_length (mms_t *this);
 void     mms_close (mms_t *this);
 
-int      mms_peek_header (mms_t *this, char *data, int maxsize);
+size_t   mms_peek_header (mms_t *this, char *data, size_t maxsize);
 
 off_t    mms_get_current_pos (mms_t *this);
