+++ src/input/mmsh.h
@@ -17,7 +17,7 @@
  * along with this program; if not, write to the Free Software
  * Foundation, Inc., 59 Temple Place - Suite 330, Boston, MA  02111-1307, USA
  *
- * $Id: mmsh.h,v 1.4 2004/04/06 00:25:29 tmattern Exp $
+ * $Id: mmsh.h,v 1.5 2006/06/20 01:46:41 dgp85 Exp $
  *
  * libmmsh public header
  */
@@ -37,7 +37,7 @@ int      mmsh_read (mmsh_t *this, char *data, int len);
 uint32_t mmsh_get_length (mmsh_t *this);
 void     mmsh_close (mmsh_t *this);
 
-int      mmsh_peek_header (mmsh_t *this, char *data, int maxsize);
+size_t   mmsh_peek_header (mmsh_t *this, char *data, size_t maxsize);
 
 off_t    mmsh_get_current_pos (mmsh_t *this);
